requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
bcrypt==3.2.0
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
distlib==0.3.1
distro==1.5.0
Django==3.1.1
django-environ==0.4.5
django-environment==0.3.8
django-filter==2.3.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-modeltranslation==0.15.1
django-mptt==0.11.0
django-postgres-extensions==0.9.3
django-reversion==3.0.8
django-translated-fields==0.10.2
django-waitress==0.1.0
django-webserver==1.2.0
djangorestframework==3.11.1
docker==4.3.1
docker-compose==1.27.0
dockerpty==0.4.1
docopt==0.6.2
drf-writable-nested==0.6.1
environ==1.0
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
jsonschema==3.2.0
Markdown==3.2.2
more-itertools==8.5.0
packaging==20.4
paramiko==2.7.2
pbr==5.5.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
pytest-docker==0.8.0
python-dotenv==0.14.0
pytz==2020.1
pyuca==1.2
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stevedore==3.2.1
texttable==1.6.3
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
waitress==1.4.4
wcwidth==0.2.5
websocket-client==0.57.0
zipp==3.1.0
